+++ src/blob/index/header.rs
@@ -1,7 +1,10 @@
+use crate::error::ValidationErrorKind;
+
 use super::prelude::*;
 
 #[derive(Debug, Deserialize, Serialize, Clone)]
 pub(crate) struct IndexHeader {
+    magic_byte: u64,
     pub records_count: usize,
     // contains serialized size of record headers, which allows to calculate offset in
     // case of `OnDisk` state of indices
@@ -11,14 +14,21 @@ pub(crate) struct IndexHeader {
     // this field also contains `written` bit (the first one)
     // to get the version, you should proceed `version >> 1`
     pub(crate) version: u8,
+    pub(crate) blob_size: u64,
 }
 
 impl IndexHeader {
-    pub fn new(record_header_size: usize, records_count: usize, meta_size: usize) -> Self {
+    pub fn new(
+        record_header_size: usize,
+        records_count: usize,
+        meta_size: usize,
+        blob_size: u64,
+    ) -> Self {
         Self {
             records_count,
             record_header_size,
             meta_size,
+            blob_size,
             ..Self::default()
         }
     }
@@ -28,12 +38,14 @@ impl IndexHeader {
         records_count: usize,
         meta_size: usize,
         hash: Vec<u8>,
+        blob_size: u64,
     ) -> Self {
         Self {
             records_count,
             record_header_size,
             meta_size,
             hash,
+            blob_size,
             ..Self::default()
         }
     }
@@ -49,6 +61,10 @@ impl IndexHeader {
 
     pub(crate) fn version(&self) -> u8 {
         self.version >> 1
+    }
+
+    pub(crate) fn magic_byte(&self) -> u64 {
+        self.magic_byte
     }
 
     #[allow(dead_code)]
@@ -71,6 +87,18 @@ impl IndexHeader {
     pub(crate) fn from_raw(buf: &[u8]) -> bincode::Result<Self> {
         bincode::deserialize(buf)
     }
+
+    pub(crate) fn validate_without_version(&self) -> Result<()> {
+        if !self.is_written() {
+            let param = ValidationErrorKind::IndexNotWritten;
+            return Err(Error::validation(param, "missing 'written' bit").into());
+        }
+        Ok(())
+    }
+
+    pub(crate) fn blob_size(&self) -> u64 {
+        self.blob_size
+    }
 }
 
 impl Default for IndexHeader {
@@ -79,8 +107,10 @@ impl Default for IndexHeader {
             records_count: 0,
             record_header_size: 0,
             meta_size: 0,
+            blob_size: 0,
             hash: vec![0; ring::digest::SHA256.output_len],
             version: HEADER_VERSION << 1,
+            magic_byte: INDEX_HEADER_MAGIC_BYTE,
         }
     }
 }