+++ src/blob/index/mod.rs
@@ -2,18 +2,19 @@ use crate::prelude::*;
 
 mod bptree;
 mod core;
-mod header;
+pub(crate) mod header;
 mod simple;
 mod tools;
 
 #[cfg(test)]
 mod benchmarks;
 
-use bptree::BPTreeFileIndex;
+pub(crate) use bptree::BPTreeFileIndex;
 use header::IndexHeader;
 
 pub(crate) use self::core::{
     FileIndexTrait, InMemoryIndex, Index, IndexConfig, MemoryAttrs, HEADER_VERSION,
+    INDEX_HEADER_MAGIC_BYTE,
 };
 pub(crate) use super::prelude::*;
 pub(crate) use crate::filter::range::RangeFilter;
@@ -30,7 +31,7 @@ pub(crate) trait IndexTrait<K>: Send + Sync {
     fn push(&mut self, h: RecordHeader) -> Result<()>;
     async fn contains_key(&self, key: &K) -> Result<bool>;
     fn count(&self) -> usize;
-    async fn dump(&mut self) -> Result<usize>;
-    async fn load(&mut self) -> Result<()>;
+    async fn dump(&mut self, blob_size: u64) -> Result<usize>;
+    async fn load(&mut self, blob_size: u64) -> Result<()>;
     fn mark_all_as_deleted(&mut self, key: &K) -> Result<Option<u64>>;
 }